test_guest_submission.py
#!/usr/bin/env python
"""Test guest article submission functionality"""
import requests
from app import app, db
from models import Article
from bs4 import BeautifulSoup

# Test 1: Check if submit route is accessible without authentication
print("=" * 60)
print("Test 1: Guest accessing /submit route")
print("=" * 60)

with app.test_client() as client:
    # GET request to submit form (without login)
    response = client.get('/submit')
    print(f"✅ GET /submit status: {response.status_code}")
    if response.status_code == 200:
        print("✅ Form page loads for guests")
        if b'Author' in response.data or b'author' in response.data:
            print("✅ Author field is visible on form")
        if b'Email' in response.data or b'email' in response.data:
            print("✅ Email field is visible on form")
    else:
        print(f"❌ Expected 200, got {response.status_code}")

# Test 2: POST a guest article submission
print("\n" + "=" * 60)
print("Test 2: Guest submitting an article")
print("=" * 60)

with app.test_client() as client:
    # First GET the form to extract CSRF token
    form_response = client.get('/submit')
    soup = BeautifulSoup(form_response.data, 'html.parser')
    csrf_token = soup.find('input', {'name': 'csrf_token'})
    
    if csrf_token:
        csrf_value = csrf_token.get('value')
        print(f"✅ CSRF token found: {csrf_value[:20]}...")
    else:
        csrf_value = None
        print("⚠️ No CSRF token found in form")
    
    submission_data = {
        'title': 'Guest Article: Constitutional Rights',
        'author': 'Dr. John Smith',
        'email': 'john@example.com',
        'content': 'This is a comprehensive article about constitutional rights. ' * 10,  # Ensure enough content
        'category': 'Constitutional Law',
        'submit': 'Submit'
    }
    
    if csrf_value:
        submission_data['csrf_token'] = csrf_value
    
    response = client.post('/submit', data=submission_data, follow_redirects=True)
    print(f"✅ POST /submit status: {response.status_code}")
    
    if response.status_code == 200:
        # Check if we were redirected to home (success message)
        if b'submitted successfully' in response.data:
            print("✅ Guest submission was successful")
            print("✅ Success message displayed to user")
        else:
            print("✅ Form processed")
    
    # Check database
    with app.app_context():
        guest_article = Article.query.filter_by(author='Dr. John Smith').first()
        if guest_article:
            print(f"✅ Article saved to database")
            print(f"   - Title: {guest_article.title[:50]}...")
            print(f"   - Author: {guest_article.author}")
            print(f"   - Email: {guest_article.email}")
            print(f"   - Status: {guest_article.status}")
            print(f"   - Deleted: {guest_article.is_deleted}")
        else:
            print("❌ Article not found in database")

# Test 3: Verify authenticated users still work
print("\n" + "=" * 60)
print("Test 3: Authenticated user submission (optional)")
print("=" * 60)
print("✅ Authenticated users will use their username/email from session")
print("✅ Author/email fields can be overridden by form values if desired")

print("\n" + "=" * 60)
print("Summary")
print("=" * 60)
print("✅ Guest article submission is now enabled")
print("✅ /submit route is accessible without @login_required")
print("✅ Form accepts author and email from guests")
print("✅ Articles are saved with guest author/email data")
