requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-Limiter==3.5.0
WTForms==3.2.1
email-validator==2.2.0
Werkzeug==3.1.3
SQLAlchemy==2.0.41
python-dotenv==1.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
Pillow==11.1.0
bleach==6.1.0
requests==2.31.0
paystack-python==0.2.0
python-slugify==8.0.1
markdown==3.5.2